pyproject.toml
[tool.poetry]
name = "legal-colombia"
version = "0.1.0"
description = ""
authors = ["jjovalle99 <jj.ovalle@uniandes.edu.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.data.dependencies]
playwright = "^1.42.0"
beautifulsoup4 = "^4.12.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
